Move ONNX golden validation under ONNX
