+++ rust/crates/photos/src/ml/mod.rs
@@ -3,12 +3,11 @@ mod clip;
 mod diagnostics;
 pub mod error;
 pub mod face;
-pub mod golden;
-mod golden_data;
-pub mod golden_tooling;
 pub mod indexing;
 mod model;
 mod onnx;
+pub use onnx::golden;
+pub use onnx::golden::tooling as golden_tooling;
 mod pet;
 mod postprocess;
 mod preprocess;
